don't tick status effects if entity is dead

+++ src/map/status_effect_container.cpp
@@ -1693,7 +1693,8 @@ void CStatusEffectContainer::TickEffects(time_point tick)
     {
         for (const auto& PStatusEffect : m_StatusEffectSet)
         {
-            if (PStatusEffect->GetTickTime() != 0 &&
+            if (!m_POwner->isDead() && // stop processing ticks if mob is dead
+                PStatusEffect->GetTickTime() != 0 &&
                 PStatusEffect->GetElapsedTickCount() <= std::chrono::duration_cast<std::chrono::milliseconds>(tick - PStatusEffect->GetStartTime()).count() / PStatusEffect->GetTickTime())
             {
                 if (PStatusEffect->GetFlag() & EFFECTFLAG_AURA)
@@ -1705,8 +1706,12 @@ void CStatusEffectContainer::TickEffects(time_point tick)
             }
         }
     }
-    DeleteStatusEffects();
-    m_POwner->PAI->EventHandler.triggerListener("EFFECTS_TICK", m_POwner);
+    // separate check after loop if mob is still alive (Could have been killed by a status effect tick)
+    if (!m_POwner->isDead())
+    {
+        DeleteStatusEffects();
+        m_POwner->PAI->EventHandler.triggerListener("EFFECTS_TICK", m_POwner);
+    }
 }
 
 /************************************************************************
